+++ src/game/ai/mod.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 mod action;
+mod ai;
 mod cost;
 mod field;
 mod orientation;
@@ -12,13 +13,15 @@ mod stone;
 mod util;
 mod visited;
 
-use action::Action;
 use cost::Cost;
-use field::Field;
 use orientation::Orientation;
 use orientation::Orientations;
 use search::actions;
 use search::search;
 use state::State;
-use stone::Stone;
 use visited::VisitedStones;
+
+pub(super) use action::Action;
+pub(super) use ai::AI;
+pub(super) use field::Field;
+pub(super) use stone::Stone;