+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ----------
 - Handle stone rotation and horizontal stone movements while pausing to
   clear complete lines
+- Switched to using Backspace key for restarting the game
 
 
 0.1.0