+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ----------
 - Fixed update of internal field state on final stone collision to
   prevent potential invariant violation
+- Added auto-playing AI
 
 
 0.1.1